+++ lib/src/service/user_service.dart
@@ -1,10 +1,11 @@
 import '../../restrr.dart';
+import 'api_service.dart';
 
 class UserService extends ApiService {
   const UserService({required super.api});
 
   Future<RestResponse<User>> login(String username, String password) async {
-    return ApiService.request(
+    return request(
         route: UserRoutes.login.compile(),
         body: {
           'username': username,
@@ -17,13 +18,13 @@ class UserService extends ApiService {
   }
 
   Future<RestResponse<bool>> logout() async {
-    return ApiService.noResponseRequest(route: UserRoutes.logout.compile(), errorMap: {
+    return noResponseRequest(route: UserRoutes.logout.compile(), errorMap: {
       401: RestrrError.notSignedIn,
     });
   }
 
   Future<RestResponse<User>> register(String username, String password, {String? email, String? displayName}) async {
-    return ApiService.request(
+    return request(
         route: UserRoutes.register.compile(),
         body: {
           'username': username,
@@ -38,11 +39,8 @@ class UserService extends ApiService {
   }
 
   Future<RestResponse<User>> getSelf() async {
-    return ApiService.request(
-        route: UserRoutes.me.compile(),
-        mapper: (json) => api.entityBuilder.buildUser(json),
-        errorMap: {
-          401: RestrrError.notSignedIn,
-        });
+    return request(route: UserRoutes.me.compile(), mapper: (json) => api.entityBuilder.buildUser(json), errorMap: {
+      401: RestrrError.notSignedIn,
+    });
   }
 }